# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface TrafficEntry {
  id: string;
  timestamp: string;
  source: string;
  method: string;
  targetUrl: string;
  finalUrl: string;
  statusCode?: number | null;
  contentType?: string | null;
  durationMs: number;
  responseSize: number;
  proxyUsed?: string | null;
  fallbackToDirect: boolean;
  error?: string | null;
  responseBodyPreview?: string | null;
}

export interface TrafficStats {
  total: number;
  errors: number;
  avgDurationMs: number;
}

export interface ProxyEntry {
  id: string;
  url: string;
  protocol: string;
  label?: string | null;
  group?: string | null;
  addedAt: string;
  lastUsedAt?: string | null;
  lastCheckedAt?: string | null;
  successCount: number;
  failureCount: number;
  consecutiveFails: number;
  alive: boolean;
  latencyMs?: number | null;
}

export interface PoolStats {
  total: number;
  alive: number;
  dead: number;
}

export interface ProxyListResponse {
  stats: PoolStats;
  proxies: ProxyEntry[];
}

export interface CheckResultEntry {
  id?: string;
  url?: string;
  alive?: boolean;
  latencyMs?: number | null;
  error?: string | null;
}

export interface CheckAllResponse {
  total: number;
  stats: PoolStats;
  results: CheckResultEntry[];
}

export interface SchedulerStatus {
  enabled: boolean;
  intervalMs: number;
  intervalMinutes: number;
  testUrl: string;
  autoClearDead?: boolean;
  lastRunAt?: string | null;
  nextRunAt?: string | null;
  lastAutoClearedCount?: number | null;
  runCount: number;
}

export interface SchedulerActionResponse {
  message: string;
  status: SchedulerStatus;
}

export type AddProxiesBodyUrlsItem = {
  url?: string;
  label?: string;
};

export type AddProxiesBody = {
  url?: string;
  label?: string;
  group?: string;
  urls?: AddProxiesBodyUrlsItem[];
};

export type AddProxies201 = { [key: string]: unknown };

export type ClearDeadProxies200 = {
  removed: number;
  message: string;
};

export type CheckAllProxiesBody = {
  testUrl?: string;
};

export type StartSchedulerBody = {
  intervalMinutes?: number;
  testUrl?: string;
  autoClearDead?: boolean;
};

export type ListTrafficParams = {
  url?: string;
  status?: string;
  source?: string;
  limit?: number;
  offset?: number;
};

export type ListTraffic200 = {
  total: number;
  entries: TrafficEntry[];
  stats: TrafficStats;
};

export type CheckProxyBody = {
  testUrl?: string;
};

export type CheckProxy200 = {
  alive: boolean;
  latencyMs?: number | null;
  statusCode?: number | null;
  error?: string | null;
};