# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: proxies
    description: Proxy pool management
  - name: scheduler
    description: Proxy health check scheduler
  - name: traffic
    description: Traffic capture log
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /proxies:
    get:
      operationId: listProxies
      tags: [proxies]
      summary: List all proxies and pool stats
      responses:
        "200":
          description: Proxy list with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyListResponse"
    post:
      operationId: addProxies
      tags: [proxies]
      summary: Add one or multiple proxies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                label:
                  type: string
                group:
                  type: string
                urls:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                      label:
                        type: string
      responses:
        "201":
          description: Proxy added
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: clearProxies
      tags: [proxies]
      summary: Clear all proxies
      responses:
        "200":
          description: All proxies cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /proxies/dead:
    delete:
      operationId: clearDeadProxies
      tags: [proxies]
      summary: Remove all dead proxies from the pool
      responses:
        "200":
          description: Dead proxies removed
          content:
            application/json:
              schema:
                type: object
                required: [removed, message]
                properties:
                  removed:
                    type: integer
                  message:
                    type: string

  /proxies/check-all:
    post:
      operationId: checkAllProxies
      tags: [proxies]
      summary: Check health of all proxies
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                testUrl:
                  type: string
      responses:
        "200":
          description: Health check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckAllResponse"

  /proxies/scheduler:
    get:
      operationId: getSchedulerStatus
      tags: [scheduler]
      summary: Get scheduler status
      responses:
        "200":
          description: Scheduler status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerStatus"

  /proxies/scheduler/start:
    post:
      operationId: startScheduler
      tags: [scheduler]
      summary: Start or reconfigure the scheduler
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                intervalMinutes:
                  type: integer
                testUrl:
                  type: string
                autoClearDead:
                  type: boolean
      responses:
        "200":
          description: Scheduler started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerActionResponse"

  /proxies/scheduler/stop:
    post:
      operationId: stopScheduler
      tags: [scheduler]
      summary: Stop the scheduler
      responses:
        "200":
          description: Scheduler stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerActionResponse"

  /proxies/scheduler/run-now:
    post:
      operationId: runSchedulerNow
      tags: [scheduler]
      summary: Manually trigger health check
      responses:
        "200":
          description: Health check triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerActionResponse"

  /traffic:
    get:
      operationId: listTraffic
      tags: [traffic]
      summary: List captured traffic entries
      parameters:
        - name: url
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Traffic list
          content:
            application/json:
              schema:
                type: object
                required: [total, entries, stats]
                properties:
                  total:
                    type: integer
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrafficEntry"
                  stats:
                    $ref: "#/components/schemas/TrafficStats"
    delete:
      operationId: clearTraffic
      tags: [traffic]
      summary: Clear all traffic entries
      responses:
        "200":
          description: Traffic cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /traffic/{id}:
    get:
      operationId: getTrafficEntry
      tags: [traffic]
      summary: Get a single traffic entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Traffic entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /proxies/{id}:
    delete:
      operationId: deleteProxy
      tags: [proxies]
      summary: Delete a proxy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proxy deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /proxies/{id}/check:
    post:
      operationId: checkProxy
      tags: [proxies]
      summary: Check health of a single proxy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                testUrl:
                  type: string
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  alive:
                    type: boolean
                  latencyMs:
                    type: integer
                    nullable: true
                  statusCode:
                    type: integer
                    nullable: true
                  error:
                    type: string
                    nullable: true
                required: [alive]
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    TrafficEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
        source:
          type: string
        method:
          type: string
        targetUrl:
          type: string
        finalUrl:
          type: string
        statusCode:
          type: integer
          nullable: true
        contentType:
          type: string
          nullable: true
        durationMs:
          type: integer
        responseSize:
          type: integer
        proxyUsed:
          type: string
          nullable: true
        fallbackToDirect:
          type: boolean
        error:
          type: string
          nullable: true
        responseBodyPreview:
          type: string
          nullable: true
      required: [id, timestamp, source, method, targetUrl, finalUrl, durationMs, responseSize, fallbackToDirect]

    TrafficStats:
      type: object
      properties:
        total:
          type: integer
        errors:
          type: integer
        avgDurationMs:
          type: integer
      required: [total, errors, avgDurationMs]

    ProxyEntry:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        protocol:
          type: string
        label:
          type: string
          nullable: true
        group:
          type: string
          nullable: true
        addedAt:
          type: string
        lastUsedAt:
          type: string
          nullable: true
        lastCheckedAt:
          type: string
          nullable: true
        successCount:
          type: integer
        failureCount:
          type: integer
        consecutiveFails:
          type: integer
        alive:
          type: boolean
        latencyMs:
          type: integer
          nullable: true
      required: [id, url, protocol, addedAt, successCount, failureCount, consecutiveFails, alive]

    PoolStats:
      type: object
      properties:
        total:
          type: integer
        alive:
          type: integer
        dead:
          type: integer
      required: [total, alive, dead]

    ProxyListResponse:
      type: object
      properties:
        stats:
          $ref: "#/components/schemas/PoolStats"
        proxies:
          type: array
          items:
            $ref: "#/components/schemas/ProxyEntry"
      required: [stats, proxies]

    CheckResultEntry:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        alive:
          type: boolean
        latencyMs:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true

    CheckAllResponse:
      type: object
      properties:
        total:
          type: integer
        stats:
          $ref: "#/components/schemas/PoolStats"
        results:
          type: array
          items:
            $ref: "#/components/schemas/CheckResultEntry"
      required: [total, stats, results]

    SchedulerStatus:
      type: object
      properties:
        enabled:
          type: boolean
        intervalMs:
          type: integer
        intervalMinutes:
          type: integer
        testUrl:
          type: string
        autoClearDead:
          type: boolean
        lastRunAt:
          type: string
          nullable: true
        nextRunAt:
          type: string
          nullable: true
        lastAutoClearedCount:
          type: integer
          nullable: true
        runCount:
          type: integer
      required: [enabled, intervalMs, intervalMinutes, testUrl, runCount]

    SchedulerActionResponse:
      type: object
      properties:
        message:
          type: string
        status:
          $ref: "#/components/schemas/SchedulerStatus"
      required: [message, status]